Add Data-ink Maximization
Remove the y-axis & all the numbers on it
Add the y-values on the glyphs (whether it is the bars or the points) as follows
For the bar charts, add the y-values on the top of the bars.
Remove the x-axis minor ticks. Only leave the major ticks.
You can use the axis orientation and range setting tricks to fix your x-axis labels.
Remove the x and y grid lines.
Remove the figure outline box.